Avoid an NPE if the abstract path does not denote a directory.
git-svn-id: https://svn.apache.org/repos/asf/felix/trunk@819679 13f79535-47bb-0310-9956-ffa450edef68

+++ ipojo/online-manipulator/src/main/java/org/apache/felix/ipojo/online/manipulator/IPOJOURLHandler.java
@@ -74,8 +74,10 @@ public class IPOJOURLHandler extends org.osgi.service.url.AbstractURLStreamHandl
      */
     public void stop() {
         File[] files = m_temp.listFiles();
-        for (int i = 0; i < files.length; i++) {
-            files[i].delete();
+        if (files != null) {
+            for (int i = 0; i < files.length; i++) {
+                files[i].delete();
+            }
         }
         m_temp.delete();
     }
